pid的参数修改为浮点型，在三个pid函数里面增加了计算pwm的时候强制类型转化为int

+++ Seekfree_TC264_Opensource_Library/CODE/PID.c
@@ -39,7 +39,7 @@ uint32 Steer_Position_PID(float SlopeBias,SteerPID K)//舵机位置式PID控制，采用分
     static float LastSlopeBias,Integral;
     int PWM;
     Integral+=SlopeBias;
-    PWM=K.P*SlopeBias+K.I*Integral+K.D*(SlopeBias-LastSlopeBias);
+    PWM=(int)(K.P*SlopeBias+K.I*Integral+K.D*(SlopeBias-LastSlopeBias));
     LastSlopeBias=SlopeBias;
     return STEER_MID+PWM;//假设斜率的范围为[-5,5]，而舵机打角PWM的范围为[850,680]，减去中值之后就能映射到[-85,85]，于此对应，所以返回值应该负号再加中值，KP先猜测为17
 }
@@ -82,7 +82,7 @@ int Speed_PI_Left(int16 left_encoder,int16 left_target,MotorPID K)
     static int Bias,Last_Bias,PWM;    //当前偏差，上一次偏差，输出的PWM
 
     Bias=left_target-left_encoder;           //求出当前偏差，期望值-当前值
-    PWM+=K.P*(Bias-Last_Bias)+K.I*Bias;      //增量式PI，并把结果直接叠加在上一次的PWM上
+    PWM+=(int)(K.P*(Bias-Last_Bias)+K.I*Bias);      //增量式PI，并把结果直接叠加在上一次的PWM上
     Last_Bias=Bias;                          //为下一次PID保存这一次偏差
 
     return PWM;         //返回可以直接赋值给电机的PWM
@@ -106,7 +106,7 @@ int Speed_PI_Right(int16 right_encoder,int16 right_target,MotorPID K)
     static int Bias,Last_Bias,PWM;    //当前偏差，上一次偏差，输出的PWM
 
     Bias=right_target-right_encoder;         //求出当前偏差，期望值-当前值
-    PWM+=K.P*(Bias-Last_Bias)+K.I*Bias;      //增量式PI，并把结果直接叠加在上一次的PWM上
+    PWM+=(int)(K.P*(Bias-Last_Bias)+K.I*Bias);      //增量式PI，并把结果直接叠加在上一次的PWM上
     Last_Bias=Bias;                          //为下一次PID保存这一次偏差
 
     return PWM;         //返回可以直接赋值给电机的PWM
